Add support of choosing position_ids in tests
