+++ models/demos/deepseek_v3/tests/test_decoder_block.py
@@ -12,6 +12,10 @@ from transformers.configuration_utils import PretrainedConfig
 import ttnn
 from models.demos.deepseek_v3.conftest import PREFILL_SEQ_LENS
 from models.demos.deepseek_v3.reference.modeling_deepseek import DeepseekV3DecoderLayer
+from models.demos.deepseek_v3.tests.pytest_utils import (
+    build_expanded_test_ids,
+    expand_test_cases_with_position_ids_ranges,
+)
 from models.demos.deepseek_v3.tt.decoder_block.decoder_block_1d import DecoderBlock1D
 from models.demos.deepseek_v3.tt.decoder_block.decoder_block_1d_base import DecoderBlock1DBase
 from models.demos.deepseek_v3.tt.decoder_block.decoder_block_2d import DecoderBlock2D
@@ -44,6 +48,7 @@ def generate_reference_io(
     batch_size: int,
     mode: str,
     state_dict: dict[str, torch.Tensor],
+    decode_position_id: int | None = None,
 ):
     reference_model = DeepseekV3DecoderLayer(hf_config, layer_idx=layer_idx).eval().to(torch.bfloat16)
     if module_path is not None:
@@ -62,7 +67,18 @@ def generate_reference_io(
     if mode == "prefill":
         position_ids_or_seq_lens = torch.tensor([seq_len])
     else:
-        position_ids = position_ids_or_seq_lens = torch.randint(0, hf_config.max_seq_len - 1, (batch_size,))
+        if decode_position_id is None:
+            position_ids = position_ids_or_seq_lens = torch.randint(
+                0, hf_config.max_seq_len - 1, (batch_size,), dtype=torch.long
+            )
+        else:
+            if not isinstance(decode_position_id, int):
+                raise ValueError(f"decode_position_id must be int or None, got {type(decode_position_id)}")
+            if not (0 <= decode_position_id < hf_config.max_seq_len):
+                raise ValueError(
+                    f"decode_position_id must be in [0, {hf_config.max_seq_len - 1}], got {decode_position_id}"
+                )
+            position_ids = position_ids_or_seq_lens = torch.ones(batch_size, dtype=torch.long) * decode_position_id
     reference_output, input_cache, output_cache = run_reference_with_attention(
         reference_model, torch_input, position_ids_or_seq_lens, layer_idx, hf_config, mode, False
     )
@@ -88,6 +104,7 @@ def run_test_forward_pass_decoder1d(
     ccl,
     force_recalculate_weight_config,
     state_dict,
+    decode_position_ids: int | None = None,
 ):
     # Check params
     if mode == "prefill":
@@ -104,6 +121,7 @@ def run_test_forward_pass_decoder1d(
         batch_size,
         mode,
         state_dict,
+        decode_position_ids,
     )
 
     # Set up page config
@@ -198,6 +216,7 @@ def run_test_forward_pass_decoder2d(
     ccl,
     force_recalculate_weight_config,
     state_dict,
+    decode_position_ids: int | None = None,
 ):
     # Check params
     if mode == "prefill":
@@ -216,6 +235,7 @@ def run_test_forward_pass_decoder2d(
         batch_size,
         mode,
         state_dict,
+        decode_position_ids,
     )
 
     # Set up page config
@@ -292,6 +312,29 @@ def run_test_forward_pass_decoder2d(
     assert_hidden_dim_pcc(tt_output_torch, reference_output, pcc_required=0.9899)
 
 
+# Base test cases - ranges will be expanded into individual test cases
+# see documentation for expand_test_cases_with_position_ids_ranges for more details
+BASE_TEST_CASES = [
+    # mode, seq_len, batch_size_per_row, decode_position_ids
+    ("decode", 1, USERS_PER_ROW, None),
+] + [
+    ("prefill", seq_len, 1, None)
+    if seq_len == 128
+    else pytest.param(
+        "prefill",
+        seq_len,
+        1,
+        None,
+        marks=pytest.mark.skip(
+            f"Skipping prefilling with seq_len={seq_len} since this would cause us to exceed our available CI workload time"
+        ),
+    )
+    for seq_len in PREFILL_SEQ_LENS
+]
+EXPANDED_TEST_CASES = expand_test_cases_with_position_ids_ranges(BASE_TEST_CASES)
+EXPANDED_TEST_IDS = build_expanded_test_ids(EXPANDED_TEST_CASES)
+
+
 @pytest.mark.parametrize(
     "device_params",
     [
@@ -353,23 +396,9 @@ def run_test_forward_pass_decoder2d(
     ],
 )
 @pytest.mark.parametrize(
-    "mode, seq_len, batch_size_per_row",
-    [
-        ("decode", 1, 32),
-    ]
-    + [
-        ("prefill", seq_len, 1)
-        if seq_len == 128
-        else pytest.param(
-            "prefill",
-            seq_len,
-            1,
-            marks=pytest.mark.skip(
-                f"Skipping prefilling with seq_len={seq_len} since this would cause us to exceed our available CI workload time"
-            ),
-        )
-        for seq_len in PREFILL_SEQ_LENS
-    ],
+    "mode, seq_len, batch_size_per_row, decode_position_ids",
+    EXPANDED_TEST_CASES,
+    ids=EXPANDED_TEST_IDS,
 )
 def test_forward_pass(
     DecoderBlockClass: type[DecoderBlock1DBase],
@@ -378,6 +407,7 @@ def test_forward_pass(
     mode,
     seq_len,
     batch_size_per_row,
+    decode_position_ids,
     hf_config_short,
     cache_path,
     mesh_device,
@@ -388,6 +418,9 @@ def test_forward_pass(
     set_deterministic_env,
     state_dict,
 ):
+    if mode != "decode":
+        decode_position_ids = None
+
     test_closure(
         DecoderBlockClass,
         module_path,
@@ -402,6 +435,7 @@ def test_forward_pass(
         ccl,
         force_recalculate_weight_config,
         state_dict,
+        decode_position_ids,
     )
