+++ models/demos/deepseek_v3/tests/test_mla.py
@@ -6,7 +6,6 @@ from pathlib import Path
 
 import pytest
 import torch
-from _pytest.mark.structures import ParameterSet
 from loguru import logger
 from transformers.configuration_utils import PretrainedConfig
 
@@ -14,6 +13,10 @@ import ttnn
 from models.common.utility_functions import comp_pcc
 from models.demos.deepseek_v3.conftest import PREFILL_SEQ_LENS
 from models.demos.deepseek_v3.reference.modeling_deepseek import DeepseekV3Attention
+from models.demos.deepseek_v3.tests.pytest_utils import (
+    build_expanded_test_ids,
+    expand_test_cases_with_position_ids_ranges,
+)
 from models.demos.deepseek_v3.tt.mla.mla1d import MLA1D
 from models.demos.deepseek_v3.tt.mla.mla2d import MLA2D
 from models.demos.deepseek_v3.utils.config_helpers import USERS_PER_ROW, sub_state_dict
@@ -32,102 +35,6 @@ from models.demos.deepseek_v3.utils.test_utils import (
 
 PCC_REQUIRED = 0.99
 PCC_REQUIRED_KVPE = 0.999
-
-
-def expand_test_cases_with_position_ids_ranges(base_cases):
-    """
-    Expand test cases where position_ids ranges are expanded into individual test cases.
-
-    Args:
-        base_cases: List of tuples (mode, seq_len, batch_size_per_row, decode_position_ids)
-            where decode_position_ids can be:
-            - None: random position_ids
-            - int: single position_id
-            - tuple(start, end): range from start to end (inclusive), step=1
-            - tuple(start, end, step): range from start to end (inclusive) with given step
-
-    Returns:
-        List of expanded test cases with individual position_ids
-
-    Examples:
-        >>> base_cases = [
-        ...     ("decode", 1, 32, None),  # Random position_ids
-        ...     ("decode", 1, 32, 1024),  # Single position_id: 1024
-        ...     ("decode", 1, 32, (4096, 4100, 2)),  # Range with step=2: 4096, 4098, 4100
-        ... ]
-        >>> expand_test_cases_with_position_ids_ranges(base_cases)
-        [
-            ("decode", 1, 32, None),
-            ("decode", 1, 32, 1024),
-            ("decode", 1, 32, 4096),
-            ("decode", 1, 32, 4098),
-            ("decode", 1, 32, 4100),
-        ]
-    """
-
-    def _process_tuple(mode, seq_len, batch_size_per_row, decode_position_ids):
-        expanded_cases = []
-        if isinstance(decode_position_ids, tuple):
-            if len(decode_position_ids) == 2:
-                # Expand range into individual position_ids with step=1
-                start, end = decode_position_ids
-                step = 1
-            elif len(decode_position_ids) == 3:
-                # Expand range into individual position_ids with given step
-                start, end, step = decode_position_ids
-                if step <= 0:
-                    raise ValueError(f"step must be > 0, got {step}")
-            else:
-                raise ValueError(
-                    f"Invalid range format: {decode_position_ids}. Expected (start, end) or (start, end, step)"
-                )
-
-            # Expand range with step
-            for pos_id in range(start, end + 1, step):
-                expanded_cases.append((mode, seq_len, batch_size_per_row, pos_id))
-        else:
-            # Keep as is (None or int)
-            expanded_cases.append((mode, seq_len, batch_size_per_row, decode_position_ids))
-
-        return expanded_cases
-
-    expanded_cases = []
-    for sample in base_cases:
-        is_pytest_param = isinstance(sample, ParameterSet)
-        if is_pytest_param:
-            mode, seq_len, batch_size_per_row, decode_position_ids = sample.values
-        else:
-            mode, seq_len, batch_size_per_row, decode_position_ids = sample
-
-        cur_cases = _process_tuple(mode, seq_len, batch_size_per_row, decode_position_ids)
-
-        if is_pytest_param:
-            cur_cases = [pytest.param(*values, marks=sample.marks, id=sample.id) for values in cur_cases]
-        expanded_cases.extend(cur_cases)
-
-    return expanded_cases
-
-
-def build_expanded_test_ids(expanded_cases):
-    """Build pytest ids for expanded test cases."""
-    expanded_ids = []
-    for val in expanded_cases:
-        if not isinstance(val, tuple):
-            expanded_ids.append(str(val))
-            continue
-
-        if isinstance(val, ParameterSet):
-            mode, seq_len, batch_size_per_row, decode_pos = val.values
-        else:
-            mode, seq_len, batch_size_per_row, decode_pos = val
-
-        if mode == "decode":
-            pos_str = decode_pos if decode_pos is not None else "random"
-            expanded_ids.append(f"mode_{mode}_seq_{seq_len}_batch_{batch_size_per_row}_pos_{pos_str}")
-        else:
-            # Prefill ignores decode_position_ids, so omit it from the ID
-            expanded_ids.append(f"mode_{mode}_seq_{seq_len}_batch_{batch_size_per_row}")
-    return expanded_ids
 
 
 def get_cache_on_host(tt_cache: ttnn.Tensor, mesh_device: ttnn.MeshDevice) -> torch.Tensor:
