+++ models/demos/deepseek_v3/tests/test_model.py
@@ -10,6 +10,10 @@ from transformers.configuration_utils import PretrainedConfig
 import ttnn
 from models.demos.deepseek_v3.conftest import PREFILL_SEQ_LENS
 from models.demos.deepseek_v3.reference.modeling_deepseek import DeepseekV3ForCausalLM
+from models.demos.deepseek_v3.tests.pytest_utils import (
+    build_expanded_test_ids,
+    expand_test_cases_with_position_ids_ranges,
+)
 from models.demos.deepseek_v3.tt.mla.mla2d import MLA2D
 from models.demos.deepseek_v3.tt.model.row_batched_model import RowBatchedModel
 from models.demos.deepseek_v3.utils.config_helpers import USERS_PER_ROW, sub_state_dict
@@ -35,6 +39,7 @@ def generate_reference_io(
     hf_config: PretrainedConfig,
     model_path: str,
     state_dict: dict[str, torch.Tensor],
+    decode_position_id: int | None = None,
 ):
     """Generate reference input and output for the given mode using either real or random weights."""
     # This needs to be disabled as deterministic way to quantize weights is not supported
@@ -69,10 +74,18 @@ def generate_reference_io(
     if mode == "prefill":
         position_ids_or_seq_lens = torch.tensor([seq_len])
     else:
-        # position_ids = torch.randint(0, hf_config.max_seq_len - 1, (batch_size,))
-        position_ids = position_ids_or_seq_lens = torch.zeros(
-            (batch_size,), dtype=torch.long
-        )  # TODO: investigate the PCC issue with real weights
+        if decode_position_id is None:
+            position_ids = position_ids_or_seq_lens = torch.randint(
+                0, hf_config.max_seq_len - 1, (batch_size,), dtype=torch.long
+            )
+        else:
+            if not isinstance(decode_position_id, int):
+                raise ValueError(f"decode_position_id must be int or None, got {type(decode_position_id)}")
+            if not (0 <= decode_position_id < hf_config.max_seq_len):
+                raise ValueError(
+                    f"decode_position_id must be in [0, {hf_config.max_seq_len - 1}], got {decode_position_id}"
+                )
+            position_ids = position_ids_or_seq_lens = torch.ones(batch_size, dtype=torch.long) * decode_position_id
 
     logger.info(
         f"Running reference model with torch_input shape: {torch_input.shape} and position_ids shape: {position_ids_or_seq_lens.shape}"
@@ -102,6 +115,7 @@ def run_test_forward_pass_dpmodel(
     ccl,
     force_recalculate_weight_config,
     state_dict,
+    decode_position_ids: int | None = None,
 ):
     # Check params
     if mode == "prefill":
@@ -114,7 +128,7 @@ def run_test_forward_pass_dpmodel(
     # Get reference IO
     logger.info("Setting up reference IO")
     state_dict, position_ids, torch_input, reference_output, input_cache, output_cache = generate_reference_io(
-        use_real_weights, mode, seq_len, batch_size, hf_config_short, model_path, state_dict
+        use_real_weights, mode, seq_len, batch_size, hf_config_short, model_path, state_dict, decode_position_ids
     )
 
     # Set up page config
@@ -187,6 +201,30 @@ def run_test_forward_pass_dpmodel(
     assert_hidden_dim_pcc(tt_output_torch, reference_output, pcc_required=0.97)
 
 
+# Base test cases - ranges will be expanded into individual test cases
+# see documentation for expand_test_cases_with_position_ids_ranges for more details
+BASE_TEST_CASES = [
+    # mode, seq_len, batch_size_per_row, decode_position_ids
+    # ("decode", 1, USERS_PER_ROW, None), # TODO: test gives low PCC for random position_ids and non-zero position_ids cases. Need to investigate why.
+    ("decode", 1, USERS_PER_ROW, 0),
+] + [
+    ("prefill", seq_len, 1, None)
+    if seq_len == 128
+    else pytest.param(
+        "prefill",
+        seq_len,
+        1,
+        None,
+        marks=pytest.mark.skip(
+            f"Skipping prefilling with seq_len={seq_len} since this would cause us to exceed our available CI workload time"
+        ),
+    )
+    for seq_len in PREFILL_SEQ_LENS
+]
+EXPANDED_TEST_CASES = expand_test_cases_with_position_ids_ranges(BASE_TEST_CASES)
+EXPANDED_TEST_IDS = build_expanded_test_ids(EXPANDED_TEST_CASES)
+
+
 @pytest.mark.parametrize(
     "device_params",
     [
@@ -199,29 +237,16 @@ def run_test_forward_pass_dpmodel(
     [True],  # Test only with real weights for now
 )
 @pytest.mark.parametrize(
-    "mode, seq_len, batch_size_per_row",
-    [
-        ("decode", 1, 32),
-    ]
-    + [
-        ("prefill", seq_len, 1)
-        if seq_len == 128
-        else pytest.param(
-            "prefill",
-            seq_len,
-            1,
-            marks=pytest.mark.skip(
-                f"Skipping prefilling with seq_len={seq_len} since this would cause us to exceed our available CI workload time"
-            ),
-        )
-        for seq_len in PREFILL_SEQ_LENS
-    ],
+    "mode, seq_len, batch_size_per_row, decode_position_ids",
+    EXPANDED_TEST_CASES,
+    ids=EXPANDED_TEST_IDS,
 )
 def test_forward_pass(
     use_real_weights,
     mode,
     seq_len,
     batch_size_per_row,
+    decode_position_ids,
     hf_config_short,
     cache_path,
     mesh_device,
@@ -233,6 +258,10 @@ def test_forward_pass(
 ):
     # Set less layers and shorter max length for the sake of testing
     hf_config_short.num_hidden_layers = 8
+
+    # Only use decode_position_ids for decode mode
+    if mode != "decode":
+        decode_position_ids = None
 
     run_test_forward_pass_dpmodel(
         use_real_weights,
@@ -246,6 +275,7 @@ def test_forward_pass(
         ccl,
         force_recalculate_weight_config,
         state_dict,
+        decode_position_ids,
     )
